# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "École Platform API",
    "description": "EdTech SaaS platform API for K-12 schools in Morocco.\n\n## Domains\n- **IAM** — Authentication, authorization, invitations, recovery, 2FA\n- **ERP** — Classes, enrollments, teacher assignments, attendance\n- **LMS** — Courses, assignments, submissions, results, content, activities, assessments\n- **Billing** — Invoices and payments\n- **COM** — Notifications, consents, parent feed\n- **AI** — Writing assistance, recommendations, KPIs\n\n## Authentication\nAll protected endpoints require a `Bearer` access token in the `Authorization` header.\nTokens are obtained via `POST /auth/login`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/activities": {
      "get": {
        "tags": [
          "lms-activities"
        ],
        "summary": "List activities",
        "description": "List available activities.\n\nShows school-specific + platform-wide activities.",
        "operationId": "list_activities_api_v1_activities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of activities",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activities/sessions": {
      "post": {
        "tags": [
          "lms-activities"
        ],
        "summary": "Start activity session",
        "description": "Start a new activity session.\n\nValidates:\n1. Activity exists (school or platform-wide)\n2. Creates session with incremented attempt_no",
        "operationId": "create_activity_session_api_v1_activities_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivitySessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created activity session",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/activities/sessions/{session_id}/complete": {
      "post": {
        "tags": [
          "lms-activities"
        ],
        "summary": "Complete activity session",
        "description": "Complete an activity session with optional score.\n\nValidates:\n1. Session exists and belongs to the student\n2. Session is in 'started' status\n3. Updates status to 'completed' and sets score",
        "operationId": "complete_activity_session_api_v1_activities_sessions__session_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivitySessionCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed session with score",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivitySessionCompleteRequest": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "title": "ActivitySessionCompleteRequest"
      },
      "ActivitySessionCreateRequest": {
        "properties": {
          "activity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Activity Id"
          }
        },
        "type": "object",
        "required": [
          "activity_id"
        ],
        "title": "ActivitySessionCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}